03_End-to-End_RAG/requirements.txt
flask==3.1.2
openai==1.107.3
PyPDF2==3.0.1
numpy==2.3.3
python-dotenv==1.1.1
werkzeug==3.1.3
python-docx==1.1.0
striprtf==0.0.26
gunicorn==21.2.0